# api/src/main/resources/swagger.yml
openapi: 3.0.1
info:
  version: ''
  title: Link API
servers:
  - url: ''
    description: This server
security:
  - oauth: []
  - basic: [ ]
...
paths:
...
  /api/measureDef:
    get:
      tags:
        - Measure Definitions
      summary: Search measure definitions loaded in Link
      operationId: searchMeasureDefinitions
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/MeasureDefinition'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Measure Definitions
      summary: Loads the measure bundle from the request into Link. Either the request body must have a FHIR Bundle, or the url parameter must be specified.
      operationId: createOrUpdateMeasureDef
      parameters:
        - name: measureId
          in: query
          description: The ID of a measure that should be loaded. The ID specified represents a URL in the system config.
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/FHIR.Bundle'
      responses:
        '200':
          description: successful response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/measureDef/{measureId}:
    delete:
      tags:
        - Measure Definitions
      summary: Loads the measure bundle from the request into Link. Either the request body must have a FHIR Bundle, or the url parameter must be specified.
      operationId: deleteMeasureDef
      parameters:
        - name: measureId
          in: path
          description: The measureId of the measure definition to be deleted
          required: true
          schema:
            type: string
      responses:
        '204':
          description: successful response with no content returned
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/{tenantId}/conceptMap:
    put:
      tags:
        - Concept Maps
      summary: Creates or updates a concept map
      operationId: createOrUpdateConceptMap
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/ConceptMap'
      responses:
        '200':
          description: successful response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags:
        - Concept Maps
      summary: Searches concept maps for the specified tenant
      operationId: searchConceptMap
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConceptMapItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/conceptMap/{id}:
    get:
      tags:
        - Concept Maps
      summary: Retrievse a specific concept map from the tenant
      operationId: getConceptMap
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: The id of the concept map to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ConceptMap'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/data/$expunge:
    delete:
      tags:
        - Data
      summary: Expunges/deletes any data that has passed the tenant's configured retention
      operationId: expungeData
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/data/$test-fhir:
    get:
      tags:
        - Data
      summary: Tests the connectivity to the tenant's configured fhir server
      description: Performs a basic integration test against the tenant's configured FHIR server. Either `patientId` or `patientIdentifier` is required.
      operationId: test
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
        - name: patientId
          in: query
          description: The logical FHIR ID of a Patient
          example: Patient/123
          schema:
            type: string
        - name: patientIdentifier
          in: query
          description: The business ID of a Patient
          example: http://some-system|321
          schema:
            type: string
        - name: measureId
          in: query
          description: The ID of the measure that should be used to determine the query method for the integration test
          example: NHSNGlycemicControlHypoglycemicInitialPopulation
          required: true
          schema:
            type: string
        - name: periodStart
          in: query
          description: The ISO formatted (FHIR) date for the reporting period's start
          example: 2022-01-01T00:00:00.000Z
          required: true
          schema:
            type: string
        - name: periodEnd
          in: query
          description: The ISO formatted (FHIR) date for the reporting period's end
          example: 2022-01-31T23:59:59.000Z
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    MeasureDefinition:
      type: object
      properties:
        id:
          type: string
        measureId:
          type: string
        bundle:
          $ref: '#/components/schemas/FHIR.Bundle'
        lastUpdated:
          type: string
...
    ConceptMap:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        contexts:
          type: array
          items:
            type: string
        conceptMap:
          $ref: '#/components/schemas/FHIR.ConceptMap'
    ConceptMapItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        contexts:
          type: array
          items:
            type: string
...
    TestResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...
    FHIR.ConceptMap:
      type: object
...
    FHIR.Bundle:
      type: object
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  securitySchemes:
    oauth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: ''
          tokenUrl: ''
          scopes: { }
    basic:
      type: http
      scheme: basic
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'